# Conflicts:
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/constant/ConsensusConstant.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/container/BlockContainer.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/manager/RoundManager.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/model/MeetingMember.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/provider/TxQueueProvider.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/task/IsolatedBlocksProcessTask.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/task/TxProcessTask.java
#	consensus-impl/poc/poc-consensus-impl/src/main/java/io/nuls/consensus/poc/utils/ConsensusTool.java
#	consensus-impl/poc/poc-consensus/src/main/java/io/nuls/consensus/poc/provider/QueueProvider.java
#	consensus-impl/src/main/java/io/nuls/consensus/cache/manager/block/BlockCacheBuffer.java
#	consensus-impl/src/main/java/io/nuls/consensus/cache/manager/block/ConfirmingBlockCacheManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/cache/manager/member/ConsensusCacheManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/cache/manager/tx/ConfirmingTxCacheManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/cache/manager/tx/OrphanTxCacheManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/cache/manager/tx/ReceivedTxCacheManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/download/DownloadCacheHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/download/DownloadProcessor.java
#	consensus-impl/src/main/java/io/nuls/consensus/download/DownloadThreadManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/download/DownloadUtils.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/BlockHashResponse.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/NodeDownloadingStatus.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/block/BifurcateProcessor.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/block/BlockHeaderChain.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/genesis/GenesisBlock.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/params/GetBlocksHashParam.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/tx/CancelDepositTransaction.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/tx/PocJoinConsensusTransaction.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/tx/RedPunishTransaction.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/tx/RegisterAgentTransaction.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/tx/YellowPunishTransaction.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/PocBlockValidatorManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/BlockConsensusValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/BlockFieldValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/BlockMaxSizeValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/HeaderContinuityValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/header/HeaderHashValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/header/HeaderPackerValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/block/header/HeaderSignValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/tx/CommissionRateValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/tx/DepositAmountValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/entity/validator/tx/RegisterAgentFieldValidator.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/GetBlocksHashRequest.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/BlockEventHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/GetBlockHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/GetBlocksHashHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/GetTxGroupHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/NewTxEventHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/SmallBlockHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/handler/TxGroupHandler.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/notice/AssembledBlockNotice.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/notice/CancelConsensusNotice.java
#	consensus-impl/src/main/java/io/nuls/consensus/event/notice/StopConsensusNotice.java
#	consensus-impl/src/main/java/io/nuls/consensus/manager/BlockManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/manager/ConsensusManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/manager/RoundManager.java
#	consensus-impl/src/main/java/io/nuls/consensus/module/impl/PocConsensusModuleBootstrap.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/impl/BlockServiceImpl.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/impl/BlockStorageService.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/impl/PocConsensusServiceImpl.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/impl/SystemServiceImpl.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/tx/JoinConsensusTxService.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/tx/RegisterAgentTxService.java
#	consensus-impl/src/main/java/io/nuls/consensus/service/tx/YellowPunishTxService.java
#	consensus-impl/src/main/java/io/nuls/consensus/thread/BlockMaintenanceThread.java
#	consensus-impl/src/main/java/io/nuls/consensus/thread/BlockPersistenceThread.java
#	consensus-impl/src/main/java/io/nuls/consensus/thread/ConsensusMeetingRunner.java
#	consensus-impl/src/main/java/io/nuls/consensus/thread/SystemMonitorThread.java
#	consensus-impl/src/main/java/io/nuls/consensus/utils/BlockBatchDownloadUtils.java
#	consensus-impl/src/main/java/io/nuls/consensus/utils/BlockInfo.java
#	consensus-impl/src/main/java/io/nuls/consensus/utils/DepositComparator.java
#	consensus-impl/src/main/java/io/nuls/consensus/utils/DistributedBlockInfoRequestUtils.java
#	consensus/src/main/java/io/nuls/consensus/entity/Consensus.java
#	consensus/src/main/java/io/nuls/consensus/entity/GetTxGroupParam.java
#	consensus/src/main/java/io/nuls/consensus/event/GetBlockRequest.java
#	consensus/src/main/java/io/nuls/consensus/event/NotFoundEvent.java
#	consensus/src/main/java/io/nuls/consensus/event/SmallBlockEvent.java
#	ledger-impl/src/main/java/io/nuls/ledger/service/impl/UtxoLedgerServiceImpl.java
#	network-impl/src/main/java/io/nuls/network/service/impl/NodesManager.java
#	network-impl/src/main/java/io/nuls/network/service/impl/netty/ClientChannelHandler.java
#	node/src/main/java/io/nuls/Bootstrap.java
#	node/src/main/resources/logback.xml

+++ network-impl/src/main/java/io/nuls/network/service/impl/netty/ServerChannelHandler.java
@@ -5,17 +5,20 @@ import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.channel.socket.SocketChannel;
+import io.nuls.core.chain.entity.Block;
 import io.nuls.core.context.NulsContext;
 import io.nuls.core.utils.log.Log;
 import io.nuls.core.utils.network.IpUtil;
 import io.nuls.core.utils.spring.lite.annotation.Autowired;
 import io.nuls.network.constant.NetworkConstant;
 import io.nuls.network.entity.Node;
-import io.nuls.network.entity.NodeGroup;
+import io.nuls.network.message.entity.GetVersionEvent;
+import io.nuls.network.message.entity.HandshakeEvent;
 import io.nuls.network.service.NetworkService;
 
 import java.net.InetAddress;
 import java.nio.ByteBuffer;
+import java.util.Map;
 
 /**
  * @author Vive
@@ -30,44 +33,36 @@ public class ServerChannelHandler extends ChannelInboundHandlerAdapter {
     public void channelRegistered(ChannelHandlerContext ctx) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
         String nodeId = IpUtil.getNodeId(channel.remoteAddress());
-//        Log.info("---------------------- server channelRegistered ------------------------- " + nodeId);
+        Log.debug("---------------------- server channelRegistered ------------------------- " + nodeId);
+
         String remoteIP = channel.remoteAddress().getHostString();
-
-        NodeGroup outGroup = networkService.getNodeGroup(NetworkConstant.NETWORK_NODE_OUT_GROUP);
-        for (Node node : outGroup.getNodes().values()) {
+        Map<String, Node> nodeMap = getNetworkService().getNodes();
+        for (Node node : nodeMap.values()) {
             if (node.getIp().equals(remoteIP)) {
-                String localIP = InetAddress.getLocalHost().getHostAddress();
-                boolean isLocalServer = IpUtil.judgeIsLocalServer(localIP, remoteIP);
-
-                if (!isLocalServer) {
-                    ctx.channel().close();
-                    return;
-                } else {
-                    getNetworkService().removeNode(node.getId());
+                if (node.getType() == Node.OUT) {
+                    String localIP = InetAddress.getLocalHost().getHostAddress();
+                    boolean isLocalServer = IpUtil.judgeIsLocalServer(localIP, remoteIP);
+                    if (!isLocalServer) {
+                        ctx.channel().close();
+                        return;
+                    } else {
+                        System.out.println("----------------sever client register each other remove node-----------------" + node.getId());
+                        getNetworkService().removeNode(node.getId());
+                    }
                 }
             }
         }
+
         // if has a node with same ip, and it's a out node, close this channel
         // if More than 10 in nodes of the same IP, close this channel
         int count = 0;
-        for (Node n : getNetworkService().getNodes().values()) {
+        for (Node n : nodeMap.values()) {
             if (n.getIp().equals(remoteIP)) {
                 count++;
-                if (count >= Node.SAME_IP_MAX_COUNT) {
+                if (count >= NetworkConstant.SAME_IP_MAX_COUNT) {
                     ctx.channel().close();
                     return;
                 }
-//
-//                if (n.getType() == Node.OUT && n.getStatus() == Node.HANDSHAKE) {
-//                    ctx.channel().close();
-//                    return;
-//                } else {
-//                    count++;
-//                    if (count == 10) {
-//                        ctx.channel().close();
-//                        return;
-//                    }
-//                }
             }
         }
     }
@@ -76,34 +71,48 @@ public class ServerChannelHandler extends ChannelInboundHandlerAdapter {
     public void channelActive(ChannelHandlerContext ctx) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
         String nodeId = IpUtil.getNodeId(channel.remoteAddress());
-        Log.info("---------------------- server Active ------------------------- " + nodeId);
+        Log.debug("---------------------- server channelActive ------------------------- " + nodeId);
+
         String channelId = ctx.channel().id().asLongText();
-        //SocketChannel channel = (SocketChannel) ctx.channel();
         NioChannelMap.add(channelId, channel);
-        Node node = new Node(Node.IN, channel.remoteAddress().getHostString(), channel.remoteAddress().getPort(), channelId);
+        Node node = new Node(channel.remoteAddress().getHostString(), channel.remoteAddress().getPort(), Node.IN);
+        node.setChannelId(channelId);
         node.setStatus(Node.CONNECT);
-        getNetworkService().addNode(node);
-//        getNetworkService().addNodeToGroup(NetworkConstant.NETWORK_NODE_IN_GROUP, node);
+        boolean success = getNetworkService().addConnNode(node);
+        if (!success) {
+            ctx.channel().close();
+            return;
+        }
+        Block bestBlock = NulsContext.getInstance().getBestBlock();
+        HandshakeEvent event = new HandshakeEvent(NetworkConstant.HANDSHAKE_SEVER_TYPE, getNetworkService().getNetworkParam().port(),
+                bestBlock.getHeader().getHeight(), bestBlock.getHeader().getHash().getDigestHex());
+        getNetworkService().sendToNode(event, nodeId, false);
     }
 
     @Override
     public void channelInactive(ChannelHandlerContext ctx) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
+        String nodeId = IpUtil.getNodeId(channel.remoteAddress());
+        Log.debug(" ---------------------- server channelInactive ------------------------- " + nodeId);
+
         String channelId = ctx.channel().id().asLongText();
         NioChannelMap.remove(channelId);
-        String nodeId = IpUtil.getNodeId(channel.remoteAddress());
-        Log.info("---------------------- server remove ------------------------- " + nodeId);
         Node node = getNetworkService().getNode(nodeId);
         if (node != null) {
             if (channelId.equals(node.getChannelId())) {
+                System.out.println("------------ sever channelInactive remove node-------------" + node.getId());
                 getNetworkService().removeNode(nodeId);
+            } else {
+                Log.debug("--------------channel id different----------------------");
+                Log.debug(node.getChannelId());
+                Log.debug(channelId);
             }
         }
     }
 
     @Override
     public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
-        Log.debug("--------------- ServerChannelHandler exceptionCaught :" + cause.getMessage(), cause);
+        Log.info("--------------- ServerChannelHandler exceptionCaught :" + cause.getMessage());
         ctx.channel().close();
     }
 
@@ -111,6 +120,7 @@ public class ServerChannelHandler extends ChannelInboundHandlerAdapter {
     public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
         SocketChannel channel = (SocketChannel) ctx.channel();
         String nodeId = IpUtil.getNodeId(channel.remoteAddress());
+//        Log.debug(" ---------------------- server channelRead ------------------------- " + nodeId);
         Node node = getNetworkService().getNode(nodeId);
         if (node != null && node.isAlive()) {
             ByteBuf buf = (ByteBuf) msg;
